+++ src/ansys/additive/core/parametric_study/parametric_study.py
@@ -479,7 +479,7 @@ class ParametricStudy:
         max_area_energy_density: float | None = None,
         iteration: int = DEFAULT_ITERATION,
         priority: int = DEFAULT_PRIORITY,
-    ):
+    ) -> int:
         """Add single bead permutations to the parametric study.
 
         Parameters
@@ -524,6 +524,11 @@ class ParametricStudy:
             Iteration number for this set of simulations.
         priority : int, default: :obj:`DEFAULT_PRIORITY <constants.DEFAULT_PRIORITY>`
             Priority for this set of simulations.
+
+        Returns
+        -------
+        int
+            Number of single bead permutations added to the parametric study.
         """  # noqa: E501
         lt = (
             layer_thicknesses
@@ -542,6 +547,7 @@ class ParametricStudy:
         )
         min_aed = min_area_energy_density or 0.0
         max_aed = max_area_energy_density or float("inf")
+        num_permutations_added = int()
         for p in laser_powers:
             for v in scan_speeds:
                 for l in lt:
@@ -593,6 +599,8 @@ class ParametricStudy:
                             self._data_frame = pd.concat(
                                 [self._data_frame, row.to_frame().T], ignore_index=True
                             )
+                            num_permutations_added += 1
+        return num_permutations_added - self._remove_duplicate_entries(overwrite=False)
 
     @save_on_return
     def generate_porosity_permutations(
@@ -616,7 +624,7 @@ class ParametricStudy:
         max_build_rate: float | None = None,
         iteration: int = DEFAULT_ITERATION,
         priority: int = DEFAULT_PRIORITY,
-    ):
+    ) -> int:
         """Add porosity permutations to the parametric study.
 
         Parameters
@@ -702,6 +710,11 @@ class ParametricStudy:
             Iteration number for this set of simulations.
         priority : int, default: :obj:`DEFAULT_PRIORITY <constants.DEFAULT_PRIORITY>`
             Priority for this set of simulations.
+
+        Returns
+        -------
+        int
+            Number of porosity permutations added to the parametric study.
         """  # noqa: E501
         lt = (
             layer_thicknesses
@@ -742,6 +755,7 @@ class ParametricStudy:
         max_ed = max_energy_density or float("inf")
         min_br = min_build_rate or 0.0
         max_br = max_build_rate or float("inf")
+        num_permutations_added = int()
         for p in laser_powers:
             for v in scan_speeds:
                 for l in lt:
@@ -811,6 +825,8 @@ class ParametricStudy:
                                                 [self._data_frame, row.to_frame().T],
                                                 ignore_index=True,
                                             )
+                                            num_permutations_added += 1
+        return num_permutations_added - self._remove_duplicate_entries(overwrite=False)
 
     @save_on_return
     def generate_microstructure_permutations(
@@ -843,7 +859,7 @@ class ParametricStudy:
         random_seed: int | None = None,
         iteration: int = DEFAULT_ITERATION,
         priority: int = DEFAULT_PRIORITY,
-    ):
+    ) -> int:
         """Add microstructure permutations to the parametric study.
 
         Parameters
@@ -982,6 +998,11 @@ class ParametricStudy:
             Iteration number for this set of simulations.
         priority : int, default: :obj:`DEFAULT_PRIORITY <constants.DEFAULT_PRIORITY>`
             Priority for this set of simulations.
+
+        Returns
+        -------
+        int
+            Number of microstructure permutations added to the parametric study.
         """  # noqa
         lt = (
             layer_thicknesses
@@ -1036,6 +1057,7 @@ class ParametricStudy:
             melt_pool_width = melt_pool_width or MicrostructureInput.DEFAULT_MELT_POOL_WIDTH
             melt_pool_depth = melt_pool_depth or MicrostructureInput.DEFAULT_MELT_POOL_DEPTH
 
+        num_permutations_added = int()
         for p in laser_powers:
             for v in scan_speeds:
                 for l in lt:
@@ -1143,6 +1165,8 @@ class ParametricStudy:
                                                 [self._data_frame, row.to_frame().T],
                                                 ignore_index=True,
                                             )
+                                            num_permutations_added += 1
+        return num_permutations_added - self._remove_duplicate_entries(overwrite=False)
 
     @save_on_return
     def update(self, summaries: list[SingleBeadSummary | PorositySummary | MicrostructureSummary]):